Frontend Projects/cookbook_machine/cookbook_generator/requirements.txt
# CAMEL AI Framework with all optional dependencies
camel-ai[all]>=0.2.16

# Additional API client dependencies
google-generativeai>=0.8.0
anthropic>=0.40.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Data processing
pydantic>=2.0.0

# Logging
rich>=13.0.0